CS457: lev comp:fix current_ hacks, 64-bit (Spork)
UnNetHack says this fixes an obscure 64-bit bug that caused a crash
while creating the Sanctum by replacing long with int; this is merged
with a similar fix for add_opvars() in case sizeof int != sizeof long,
so I can't tell where it is.  I don't even have a 64-bit CPU, so I'll
have to take these fixes on faith.

The current_region/current_coord hacks in the level compiler grammar
have been replaced with the use of Yacc features.

This changeset originally included a fix for ROOM/SUBROOM emitting
opcodes for IF statements, but I already rolled that into commit
6fc8caae1a (lev comp: value stack, chances, etc.).

According to UnNetHack, this is from SporkHack r590-592; r588 was
mentioned, but it's the ROOM/SUBROOM IF opcode fix.

+++ libnitrohack/util/lev_main.c
@@ -566,8 +566,8 @@ void scan_map(char *map, sp_lev *sp)
 {
 	int i, len;
 	char *s1, *s2;
-	int max_len = 0;
-	int max_hig = 0;
+	long max_len = 0;
+	long max_hig = 0;
 	char msg[256];
 	char *tmpmap[ROWNO];
 	int dx, dy;
@@ -607,7 +607,7 @@ void scan_map(char *map, sp_lev *sp)
 		for (i=0; i<len; i++)
 		  if ((tmpmap[max_hig][i] = what_map_char(map[i])) == INVALID_TYPE) {
 		      sprintf(msg,
-			 "Invalid character @ (%d, %d) - replacing with stone",
+			 "Invalid character @ (%ld, %d) - replacing with stone",
 			      max_hig, i);
 		      yywarning(msg);
 		      tmpmap[max_hig][i] = STONE;
